Initial commit
First commit of changes for what will be the Beta 3 Update 3. This commit brings Better Explorer into a useable state, using the newer versions of the C++ projects from a later commit in the master branch.

The point of this branch is to only provide fixes for serious bugs that would hinder the average use of the features in this new version. No new features will be added beyond that's currently here.